[NFC][SSAF] Move EntityPointerLevel to a separate folder
EntityPointerLevel will later be shared with other summaries besides
UnsafeBufferUsage. This commit moves it to a separate file.
